requirements.txt
pydbus
pyyaml
httpx
click
rich
